Extract R8 DCE verification into featured-shrinker-tests module
Move R8EliminationTest from featured-gradle-plugin into a dedicated
non-published JVM module. Split the monolithic test (~660 lines) into
focused units: bytecode generators, JAR assembly, ProGuard rules,
test harness, assertions, and two test classes (boolean/int flags).
