+++ .gitignore
@@ -1,3 +1,4 @@
+swarm-report/
 *.iml
 .kotlin
 .gradle
