+++ featured-gradle-plugin/build.gradle.kts
@@ -57,6 +57,4 @@ mavenPublishing {
 dependencies {
     testImplementation(gradleTestKit())
     testImplementation(libs.kotlin.testJunit)
-    testImplementation(libs.r8)
-    testImplementation(libs.asm)
 }
