+++ featured-shrinker-tests/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    alias(libs.plugins.kotlinJvm)
+}
+
+kotlin {
+    jvmToolchain(21)
+}
+
+dependencies {
+    testImplementation(libs.kotlin.testJunit)
+    testImplementation(libs.r8)
+    testImplementation(libs.asm)
+}
