+++ mkdocs.yml
@@ -60,6 +60,7 @@ nav:
       - JVM: guides/jvm.md
       - Providers: guides/providers.md
       - Best Practices: guides/best-practices.md
+      - R8 DCE Verification: guides/r8-verification.md
       - iOS Dead-Code Elimination: ios-integration.md
   - API Reference: api/index.md
   - Changelog: changelog.md
