+++ settings.gradle.kts
@@ -48,6 +48,7 @@ include(":featured-platform")
 include(":featured-bom")
 include(":featured-detekt-rules")
 include(":featured-lint-rules")
+include(":featured-shrinker-tests")
 
 include(":providers:configcat")
 include(":providers:datastore")
